Grafica una red con un layout circular

Utilizando un objeto resultante de la función crear_redes_de_palabras(), crea un gráfico de redes con una distribución espacial circular. Para esto, se usan bajo la capa la librería igraph.

Uso

grafico_red_circular(red, vertex.cex = 2L)

Argumentos

red

Un objeto de clase igraph resultante de la función crear_redes_de_palabras() o de procesos similares usando el paquete igraph.

vertex.cex

multiplicador de tamaño de los nodos. Por defecto es 2.