Análisis de comunidades usando algoritmo optimization 'Greedy'

Utilizando un objeto resultante de la función crear_redes_de_palabras(), crea un análisis de comunidades usando el algoritmo de optimización 'greedy'. Para esto, se usan bajo la capa la librería igraph.

Uso

red_comunidades_greedy(red, seed = 12345, output = "plot")

Argumentos

red

Un objeto de clase igraph resultante de la función crear_redes_de_palabras() o de procesos similares usando el paquete igraph.

seed

Semilla de reproducibilidad. Número entero usando para la generación de resultados reproducibles usando set.seed() por debajo.

output

Resultado final. Puede ser 'plot' (devuelve un gráfico de análisis de comunidades), 'groups' (un vector con los grupos) o 'raw' (el objeto sin procesar de clase 'communities').