Crea redes usando algún vector de palabras u oraciones para ser analizadas

Utilizando la librería tm y igraph, opta por un 'wrapper' para generar redes listas para ser utilizadas bajo el contexto de procesamiento de lenguaje natural.

crear_redes_de_palabras(x)

Arguments

x
Un vector de carácteres.